feat: 검색 페이지에서 SearchMainContainer로 한 단계 더 분리, force-dynamic 설정 추가
